refactor(scroll-view): move bounce-related code from common to scroll-view package
refactor(feed-list): move refresh-related code from common to feed-list package
docs: update build instructions in AGENTS.md
chore: remove unused typedoc hooks generation code
