+++ packages/lynx-ui-feed-list/src/types/index.docs.ts
@@ -5,11 +5,9 @@
 import type { ForwardedRef, ReactElement } from '@lynx-js/react'
 
 import type { BaseGesture } from '@lynx-js/gesture-runtime'
-import type {
-  BounceableBasicProps,
-  RefreshProps,
-} from '@lynx-js/lynx-ui-common'
 import type { ListProps, ListRef } from '@lynx-js/lynx-ui-list'
+
+import type { BounceableBasicProps, RefreshProps } from '../hooks/useRefresh'
 
 export type FeedList = (props: FeedListProps) => ReactElement
 
@@ -78,4 +76,37 @@ export interface FeedListProps extends Omit<ListProps, 'ref'> {
    * @zh 如果您想使用手势，请在此处传递。
    */
   'main-thread:gesture'?: BaseGesture
+}
+
+export interface RefreshEvent {
+  triggeredBy: 'startRefresh' | 'drag'
+}
+
+export interface refreshOffsetEvent {
+  offset: number
+  headerSize: number
+  isDragging: boolean
+}
+
+export interface headerReleased {
+  offset: number
+  headerSize: number
+}
+
+export enum RefreshState {
+  IDLE = 0, // Hide
+  OVER_DRAG_RELEASE = 1, // Drag over refresh header size and release
+  REFRESHING = 2, // During refresh
+  DRAGGING = 3, // Dragging
+}
+
+export interface RefreshStateChange {
+  /**
+   * Status of refreshHeader
+   * @zh 刷新头的状态
+   * @Android
+   * @iOS
+   * @Harmony
+   */
+  state: RefreshState
 }
